chore(idp): Rename cmd/idp to cmd/influx
