+++ Makefile
@@ -26,7 +26,7 @@ SOURCES := $(shell find . -name '*.go' -not -name '*_test.go')
 SOURCES_NO_VENDOR := $(shell find . -path ./vendor -prune -o -name "*.go" -not -name '*_test.go' -print)
 
 # List of binary cmds to build
-CMDS := bin/idp bin/idpd bin/ifqld
+CMDS := bin/influx bin/idpd bin/ifqld
 
 # List of utilities to build as part of the build process
 UTILS := bin/pigeon bin/cmpgen bin/goreleaser
@@ -49,8 +49,8 @@ subdirs: $(SUBDIRS)
 bin/ifqld: $(SOURCES)
 	$(GO_BUILD) -i -o bin/ifqld ./cmd/ifqld
 
-bin/idp: $(SOURCES)
-	$(GO_BUILD) -i -o bin/idp ./cmd/idp
+bin/influx: $(SOURCES)
+	$(GO_BUILD) -i -o bin/influx ./cmd/influx
 
 bin/idpd: $(SOURCES)
 	$(GO_BUILD) -i -o bin/idpd ./cmd/idpd
