+++ cmd/influx/bucket.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/influxdata/platform"
-	"github.com/influxdata/platform/cmd/idp/internal"
+	"github.com/influxdata/platform/cmd/influx/internal"
 	"github.com/influxdata/platform/http"
 	"github.com/spf13/cobra"
 )
